+++ tests/it/main.rs
@@ -14,5 +14,5 @@
 
 mod layer;
 mod ops;
+mod readers;
 mod services;
-mod wraps;
